requirements.txt
guacamole-api-wrapper
openstacksdk
requests
pyyaml
jinja2
colorama
munch